pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "spyprot>=0.9.6",
    "streamlit>=1.49.1",
]
